requirements.txt
streamlit
pandas
tiktoken
openai
python-dotenv
graphrag
lancedb
langchain
langchain-community
langchain-openai
chromadb
